Correct the documentation for the number of thread parameters.
Note that this still requires more work, as it should use the available number
of threads by default. Also, it seems that setting that parameter to values
greater than zero only works as long as there are enough threads in the global
IlmThread  pool. Otherwise everything will run from the same thread
regardless of the value of this param.